pyproject.toml
[project]
name = "smu-lecture-enroller"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.5",
    "lxml>=6.0.1",
    "pillow>=11.3.0",
    "pycryptodome>=3.23.0",
    "pynput>=1.8.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "stdiomask>=0.0.6",
]
